sync implemented front end user profile

+++ GADATA/GLUE/Tables/vultijd_data.sql
@@ -4,7 +4,9 @@
     [Str1]       INT          NULL,
     [Str2]       INT          NULL,
     [Str3]       INT          NULL,
-    [ID]         INT          IDENTITY (1, 1) NOT NULL,
+    [ID]         BIGINT       IDENTITY (1, 1) NOT NULL,
     CONSTRAINT [PK_vultijd_data] PRIMARY KEY CLUSTERED ([ID] ASC)
 );
+
+
